Use `conda-incubator/setup-miniconda` for conda builds
This fixes a failure to build on Windows. This failure was inexplicable,
but introduced in a mid-November update to the GitHub actions Windows
image.

It's not clear why this fixes the issue, but it's a nice simplification
to use an existing action instead of rolling our own conda installation.

The shell specification `shell: cmd /C CALL {0}` on Windows is
necessary, otherwise conda is not activated properly in the step.
Unclear whether the `bash -l {0}` is needed, but it is more conda-like
for the shell to run as a login shell.

Generally automating conda stuff is fragile due to it relying so much on
shell manipulation, so hopefully we can defer most of the trickiness to
simply using `conda-incubator/setup-miniconda` and following its advice.