Do setuptools-conda build in ${{ runner.temp }}\cb
on Windows. Results in shorter filepaths that hopefully will keep the
builds happy. Otherwise when conda-build is setting up the build
environment (setting environment variables before entering the
developer command prompt), we see errors like:

```
The input line is too long.
The syntax of the command is incorrect.
```

This will hopefully fix the problem.