Drop building of conda packages for win-x86 platform.
Either miniconda is not available or setuptools/setuptools-scm not
available from default channels at high enough version for our builds.

This platform should get deprecated away soon-ish.
If you are reading this and need win-x86 packages,
use pip packages for now and contact the devs for options to move off
this platform.